int[] -> byte[], saves 18 precious bytes! :p

+++ DigimarcPIN/MainForm.cs
@@ -12,7 +12,7 @@ namespace DigimarcPIN
 {
     public partial class MainForm : Form
     {
-        private int[] key = { 0x1F, 0x05, 0x35, 0x07, 0x2B, 0x2F };
+        private byte[] key = { 0x1F, 0x05, 0x35, 0x07, 0x2B, 0x2F };
         // For some reason there's 8 more bytes in the key, after the 6 above.
         // If anyone has seen an ID with more than 6 digits, please let me know!
         // Complete key: 1F 05 35 07 2B 2F 0D 02 25 13 03 17 11 29
@@ -31,7 +31,7 @@ namespace DigimarcPIN
                     byte digit = (byte)(stringBytes[i] - 0x30);
                     n += digit * key[5 - i];
                 }
-                n = n % 89 + 10;
+                n = n % 89 + 10; // Probably originally intended to be % 90 instead of 89 (Digimarc's bug, not mine)
                 pinBox.Text = n.ToString();
             } else {
                 pinBox.Text = "--";
